vuln-fix: Zip Slip Vulnerability
This fixes a Zip-Slip vulnerability.

This change does one of two things. This change either

1. Inserts a guard to protect against Zip Slip.
OR
2. Replaces `dir.getCanonicalPath().startsWith(parent.getCanonicalPath())`, which is vulnerable to partial path traversal attacks, with the more secure `dir.getCanonicalFile().toPath().startsWith(parent.getCanonicalFile().toPath())`.

For number 2, consider `"/usr/outnot".startsWith("/usr/out")`.
The check is bypassed although `/outnot` is not under the `/out` directory.
It's important to understand that the terminating slash may be removed when using various `String` representations of the `File` object.
For example, on Linux, `println(new File("/var"))` will print `/var`, but `println(new File("/var", "/")` will print `/var/`;
however, `println(new File("/var", "/").getCanonicalPath())` will print `/var`.

Weakness: CWE-22: Improper Limitation of a Pathname to a Restricted Directory ('Path Traversal')
Severity: High
CVSSS: 7.4
Detection: CodeQL (https://codeql.github.com/codeql-query-help/java/java-zipslip/) & OpenRewrite (https://public.moderne.io/recipes/org.openrewrite.java.security.ZipSlip)

Reported-by: Jonathan Leitschuh <Jonathan.Leitschuh@gmail.com>

Bug-tracker: https://github.com/JLLeitschuh/security-research/issues/16

+++ src/main/java/fr/lebonq/minecraft/libraries/LibrarieNative.java
@@ -51,10 +51,13 @@ public class LibrarieNative extends Librarie {
                     if(vEntry.isDirectory()){//Si c'est un directory on recupere tout dedans avant d'aller plus loins
                     }
                     else{
-                        File vFileOut = new File(pBin + "/" +vEntry.getName());
+                        File vFileOut = new File(pBin,vEntry.getName());
+                        if(!vFileOut.toPath().normalize().startsWith(pBin.toPath().normalize())) {
+                            throw new IOException("Bad zip entry");
+                        }
                         vFileOut.deleteOnExit();
                         if(vEntry.getName().lastIndexOf("/") != -1){//Si == -1 alors c'est un fichier sans dossier
-                            File vFileFolder = new File(pBin + "/" + vEntry.getName().substring(0, vEntry.getName().lastIndexOf("/")));//to create folder without file name
+                            File vFileFolder = new File(pBin,vEntry.getName().substring(0,vEntry.getName().lastIndexOf("/")));//to create folder without file name
                             vFileFolder.mkdirs();
                             vFileFolder.deleteOnExit();//On supprimer le fichier a la du process launcher
                         }
